% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot_function.R
\name{draw_volcano_plot}
\alias{draw_volcano_plot}
\title{Title Draw volcano plot}
\usage{
draw_volcano_plot(
  file_for_analysis,
  logfc_limit = 2.5,
  fdr_limit = -log10(0.05),
  known_protein_names = NULL
)
}
\arguments{
\item{file_for_analysis}{a dataframe - output of perseus_FDR_ttest or a dataframe with at least 2 columns: logFC and FDR}

\item{logfc_limit}{- a threshold for logFC to draw a line on a plot}

\item{fdr_limit}{- a threshold for FDR to draw a line on a plot}

\item{known_protein_names}{- an optional gene names for Arabidopsis AGIs}
}
\value{
a ggplot object - volcano plot
}
\description{
Draw a volcano plot for the proteomics dataset based on logFC and FDR values
}
\author{
Evgeniya Pukhovaya
}
