% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFC_values_arabidopsis.R
\docType{data}
\name{logFC_values_arabidopsis}
\alias{logFC_values_arabidopsis}
\title{Sample Arabidopsis data with logFC and p}
\format{
a dataframe from a Perseus output table
}
\source{
Evgeniya Pukhovaya, unpublished data
}
\usage{
logFC_values_arabidopsis
}
\description{
Sample dataset of logFC and p values with Arabidopsis gene IDs
}
\keyword{datasets}
