% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFC_values_marchantia.R
\docType{data}
\name{logFC_values_marchantia}
\alias{logFC_values_marchantia}
\title{Sample Marchantia data with logFC and p}
\format{
a dataframe from a Perseus output table
}
\source{
Evgeniya Pukhovaya, unpublished data
}
\usage{
logFC_values_marchantia
}
\description{
Sample dataset of logFC and p values without any species-specific gene IDs
}
\keyword{datasets}
