% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perseus_ttest_file_preprocessing.R
\name{perseus_FDR_ttest}
\alias{perseus_FDR_ttest}
\title{Title Calculate FDR based on the list of p-values}
\usage{
perseus_FDR_ttest(file)
}
\arguments{
\item{file}{A Perseus putput dataframe. Should have at least 3 columns named: Fasta.headers, Student.s.T.test.Difference, Student.s.T.test.p.value. For Arabidopsis datasets, the column ENSG containing gene AGIs is optional}
}
\value{
a fataframe with 4 or 5 columns: minus_log_p_val = -log10(p); logFC = log2(fold change); FDR calculated by BH method; Fasta_header from Uniprot; AGI (optionally)
}
\description{
Calcualates BH FDR from the file in the Perseus output format. The input should already contain fold change and p-values
}
\author{
Evgeniya Pukhovaya
}
