% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perseus_ttest_significant.R
\name{perseus_ttest_significant}
\alias{perseus_ttest_significant}
\title{Select statisticaly significant hits}
\usage{
perseus_ttest_significant(file, logfc_threshold = 0, FDR_threshold = 0.1)
}
\arguments{
\item{file}{an output of perseus_FDR_ttest function or a dataframe with at least three columns (FDR, logFC, gene/protein ID)}

\item{logfc_threshold}{a fold change threshold}

\item{FDR_threshold}{a fold discovery rate threshold}
}
\value{
a dataframe with the same amount of columns as input filtered according to logFC and FDR thresholds
}
\description{
Filters the dataframe according to two threshold values: FDR < FDR_threshold and logFC > logFC_threshold
}
\author{
Evgeniya Pukhovaya
}
