% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins_name_arabidopsis_example.R
\docType{data}
\name{proteins_name_arabidopsis_example}
\alias{proteins_name_arabidopsis_example}
\title{Sample Arabidopsis protein name file}
\format{
a dataframe with two columns (AGI, Trivial)
}
\source{
Araport; Evgeniya Pukhovaya, unpublished data
}
\usage{
proteins_name_arabidopsis_example
}
\description{
Contains trivial gene (protein) names correspoding to the gene IDs of interest
}
\keyword{datasets}
